move decorators to file, add decorator for retrospective, log aws requests
