+++ geoglows/_constants.py
@@ -1,5 +1,10 @@
 import os
 
+ODP_CORE_S3_BUCKET_URI = 's3://geoglows-v2'
+ODP_FORECAST_S3_BUCKET_URI = 's3://geoglows-v2-forecasts'
+ODP_RETROSPECTIVE_S3_BUCKET_URI = 's3://geoglows-v2-retrospective'
+ODP_S3_BUCKET_REGION = 'us-west-2'
+
 METADATA_TABLE_PATH = os.getenv(
     'PYGEOGLOWS_METADATA_TABLE_PATH',
     os.path.join(os.path.dirname(__file__), 'data', 'metadata-tables.parquet')
